Move regex_onig_line_proc function to impure/onig module
