+++ src/main.rs
@@ -25,7 +25,7 @@ use onig::{self};
 use token::Token;
 
 const CMD: &'static str = env!("CARGO_PKG_NAME"); // "teip"
-const DEFAULT_CAP: usize = 1024;
+pub const DEFAULT_CAP: usize = 1024;
 
 pub fn msg_error(msg: &str) {
     eprintln!("{}: {}", CMD, msg);
@@ -460,7 +460,7 @@ fn main() {
                 .unwrap_or_else(|e| error_exit(&e.to_string()));
         } else if flag_regex {
             if flag_onig {
-                regex_onig_line_proc(&mut ch, &regex_onig, flag_invert, line_end)
+                impure::onig::regex_onig_line_proc(&mut ch, &regex_onig, flag_invert, line_end)
                     .unwrap_or_else(|e| error_exit(&e.to_string()));
             } else {
                 regex_line_proc(&mut ch, &regex, flag_invert, line_end)
@@ -536,47 +536,6 @@ fn line_line_proc(
             Err(e) => msg_error(&e.to_string()),
         }
         i += 1;
-    }
-    Ok(())
-}
-
-fn regex_onig_line_proc(
-    ch: &mut PipeIntercepter,
-    re: &onig::Regex,
-    invert: bool,
-    line_end: u8,
-) -> Result<(), errors::TokenSendError> {
-    let stdin = io::stdin();
-    loop {
-        let mut buf = Vec::with_capacity(DEFAULT_CAP);
-        match stdin.lock().read_until(line_end, &mut buf) {
-            Ok(n) => {
-                let eol = trim_eol(&mut buf);
-                if n == 0 {
-                    ch.send_eof()?;
-                    break;
-                }
-                let line = String::from_utf8_lossy(&buf).to_string();
-                match re.find(&line) {
-                    Some(_) => {
-                        if invert {
-                            ch.send_msg(line.to_string())?;
-                        } else {
-                            ch.send_pipe(line.to_string())?;
-                        }
-                    },
-                    None => {
-                        if invert {
-                            ch.send_pipe(line.to_string())?;
-                        } else {
-                            ch.send_msg(line.to_string())?;
-                        }
-                    }
-                };
-                ch.send_msg(eol)?;
-            }
-            Err(e) => msg_error(&e.to_string()),
-        }
     }
     Ok(())
 }
@@ -782,7 +741,7 @@ fn field_proc(
     Ok(())
 }
 
-fn trim_eol(buf: &mut Vec<u8>) -> String {
+pub fn trim_eol(buf: &mut Vec<u8>) -> String {
     if buf.ends_with(&[b'\r', b'\n']) {
         buf.pop();
         buf.pop();
